requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1

transformers==4.44.2
trl==0.10.1
diffusers==0.30.3
accelerate==0.34.2
open-clip-torch==2.26.1
ImageReward==1.5
opencv-python==4.10.0.84
moviepy==2.1.1
ffmpeg-python==0.2.0
sentencepiece==0.2.0
Pillow==10.4.0
numpy==1.26.4
openai==1.61.0
requests==2.32.3
huggingface_hub==0.24.6

safetensors==0.4.5
einops==0.8.0
protobuf==5.28.0
modelscope==1.16.1
ftfy==6.2.3
pynvml==11.5.3
pandas==2.2.2
peft==0.12.0
tqdm==4.66.5
imageio==2.35.1
imageio[ffmpeg]==2.35.1